==> notebooks/01_AM_Transform_Covid_Tweets.ipynb <==
import pandas as pd
import numpy as np
import reverse_geocode

# [OUTDATED!] Transform overall COVID-19 tweets

- add country based on coordinates
- store as csv

hydrated_json = pd.read_json("../data/raw/Hydrated_Tweets.jsonl", lines=True)

Drop faulty rows (i.e. rows without coordinates):

hydrated_json = hydrated_json.drop(hydrated_json[hydrated_json['coordinates'].isnull()].index).reset_index(drop=True)

Add country name based on coordinates:

hydrated_json["country"] = ""
for i in range(len(hydrated_json)):
    hydrated_json["country"][i] = reverse_geocode.search(tuple([hydrated_json["geo"][i]["coordinates"],(1,1)]))[0]['country'] #second tuple is to avoid error

len(hydrated_json["country"].unique())

Export dataset as CSV:

hydrated_json.to_csv("../data/interim/Hydrated_Tweets_with_countries.csv")
