==> notebooks/02_AM_Transform_Vaccine_Tweets.ipynb <==
import pandas as pd
import numpy as np

# Prepare dataset for hydration

Hydration describes the process of fetching tweet information via the twitter api. For that, we stripped the pre-hydrated dataset of everything so we can get fresh data.

#raw_df = pd.read_csv("../data/raw/kaggle_dataset.csv")

#raw_df = raw_df["id"].reset_index(drop=True)

#raw_df.to_csv("../data/raw/tweet_ids.csv", index=False)

# Transform Raw Vaccine Tweets

raw_vaccine_tweets = pd.read_json("../data/raw/vaccine_tweets_hydrated.jsonl", lines=True, encoding="iso-8859-1")

raw_vaccine_tweets

# Data preparation

- Remove duplicate tweets
    - Drop if retweeted == true
    - Remove duplicate text (or tweet ids)
- Relevant columns = id, created_at, username, full_text, retweet, hashtags

## Remove duplicate tweets

raw_vaccine_tweets[raw_vaccine_tweets.retweeted == True]

raw_vaccine_tweets.id.count()

raw_vaccine_tweets = raw_vaccine_tweets.drop_duplicates(subset=["full_text"], keep='first').reset_index(drop=True)

raw_vaccine_tweets.id.count()

## Select relevant columns

raw_vaccine_tweets.columns

raw_vaccine_tweets = raw_vaccine_tweets[["id_str","created_at","user","geo","full_text", "entities"]]

##  Extracting user_ids 

raw_vaccine_tweets["user_id"] = int

for i in range(len(raw_vaccine_tweets)):
    raw_vaccine_tweets.user_id[i] = raw_vaccine_tweets.user[i]["id"]

raw_vaccine_tweets.user_id

## Hashtags

renaming entities column to hashtags:

raw_vaccine_tweets = raw_vaccine_tweets.rename(columns={'entities': 'hashtags', 'id_str':'id'})

Extracting hashtags which are stored in entities>hashtags>text:

for i in range(len(raw_vaccine_tweets)):
    try:
        raw_vaccine_tweets.hashtags[i] = [value["text"] for value in raw_vaccine_tweets.iloc[i]["hashtags"]["hashtags"]]
    except:
        print("failed: ",i)
        

Storing hashtags:count as key:value in a dict:

hashtag_dict = {}
for i in range(len(raw_vaccine_tweets)):
    for hashtag in raw_vaccine_tweets["hashtags"][i]:
        if hashtag not in hashtag_dict:
            hashtag_dict[hashtag] = 1
        else:
            hashtag_dict[hashtag] += 1

Identifying relevant hashtags for each vaccine manufacturer:

sorted(hashtag_dict.items(), key=lambda x: x[1], reverse=True)

pfizer_biontech_vax = ["Pfizer", "PFIZER", "PfizerBioNTech", "PfizerVaccine", "pfizer", "PfizerBiontech", "BioNTech", "pfizerbiontech", "PfizerBioNtech", "Biontech", "biontech", "PFIZERBIONTECH", "BioNTechpfizer"]
sputnik_vax = ["SputnikV", "Sputnik", "Sputnikv", "sputnikv", "SputnikUpdates", "Sputnikvaccine", "sputnikV", "sputnik", "SPUTNIKV", "SputnikVaccinated", "SputnikLight", "SputnikVaccineInKenya", "SputnikVaccine"]
sinopharm_vax = ["Sinopharm", "sinopharm", "SinoPharm", "BoycottSinopharm", "SINOPHARM"]
sinovac_vax = ["Sinovac", "sinovac", "SinoVac", "SINOVAC", "BoycottSinovac"]
moderna_vax = ["Moderna", "moderna", "modernavaccine", "MODERNA", "ModernaVaccine", "modernagang", "teammoderna", "modeRNA", "ModernaGang"]
oxford_az_vax = ["OxfordAstraZeneca", "oxfordastrazeneca", "Oxford", "oxfordvaccine", "OxfordVaccine", "OxfordAstrazeneca", "AstraZeneca", "astrazeneca", "AstraZenaca", "astrazenecavaccine", "Astrazeneca", "AstraZeneka", "Astrazenaca", "ASTRAZENECA", "AstraZenecaVaccine"]
covaxin_vax = ["Covaxin", "COVAXIN", "covaxin", "GurgaonCOVAXIN", "MumbaiCOVAXIN", "covaxine", "BBMPCOVAXIN", "PuneCOVAXIN", "CoVaxin", "ThaneCOVAXIN", "covaxinated", "covaxinvaccine", "BharatBiotech", "AatmanirbharBharat", "AtmaNirbharBharat", "bharatbiotech", "bharatBiotech", "AtmanirbharBharat", "Bharat", "congressmuktbharat", "atmanirbharbharat"]
jandj_vax = ["johnsonandjohnson", "JohnsonandJohnson", "JohnsonAndJohnson", "JohnsonAndJohnsonVaccine", "Johnson", "JandJ", "JohnsonJohnson", "johnson", "JJ"]

normalizing hashtags to all lowercase:

for i in range(0, len(raw_vaccine_tweets)):
    for j in range(len(raw_vaccine_tweets["hashtags"][i])):
        review = raw_vaccine_tweets["hashtags"][i][j]
        review = review.lower()
        raw_vaccine_tweets["hashtags"][i][j] = review

adding columns for each vaccine manufacturer, based on the hashtags of a tweet:

raw_vaccine_tweets["PfizerBiontech"] = 0
raw_vaccine_tweets["SputnikV"] = 0
raw_vaccine_tweets["Sinopharm"] = 0
raw_vaccine_tweets["Sinovac"] = 0
raw_vaccine_tweets["Moderna"] = 0
raw_vaccine_tweets["AstraZeneca"] = 0
raw_vaccine_tweets["Covaxin"] = 0
raw_vaccine_tweets["JandJ"] = 0

for i in range(len(raw_vaccine_tweets)):
    for hashtag in raw_vaccine_tweets["hashtags"][i]:
        if hashtag in pfizer_biontech_vax:
            raw_vaccine_tweets["PfizerBiontech"][i] = 1
        if hashtag in sputnik_vax:
            raw_vaccine_tweets["SputnikV"][i] = 1
        if hashtag in sinopharm_vax:
            raw_vaccine_tweets["Sinopharm"][i] = 1
        if hashtag in sinovac_vax:
            raw_vaccine_tweets["Sinovac"][i] = 1
        if hashtag in moderna_vax:
            raw_vaccine_tweets["Moderna"][i] = 1
        if hashtag in oxford_az_vax:
            raw_vaccine_tweets["AstraZeneca"][i] = 1
        if hashtag in covaxin_vax:
            raw_vaccine_tweets["Covaxin"][i] = 1
        if hashtag in jandj_vax:
            raw_vaccine_tweets["JandJ"][i] = 1

---

from geopy import Nominatim
import reverse_geocode

## Geo location

- Fetching location from tweet geo data
- Engineer location from user profile
- transform both to standardized country names (needed for pyplot)

Fetching location specified in user profile:

raw_vaccine_tweets["user_location"] = None
for i in range(len(raw_vaccine_tweets["user"])):
    raw_vaccine_tweets["user_location"][i] = raw_vaccine_tweets["user"][i]["location"]

countries = ['argentina', 'australia', 'austria', 'belgium', 'brazil', 'canada', 'france', 'germany', 'india', 'israel', 'italy', 'japan', 'mexico', 'pakistan', 'russia', 'spain', 'uae', 'uk', 'usa']

Extract location from user profile and standardize country name:

#set country for locations that contain that country
for country in countries :
    raw_vaccine_tweets["user_location"][raw_vaccine_tweets["user_location"].str.lower().str.contains(country)] = country

#remove any location that isn't in country-list
for i in range(len(raw_vaccine_tweets)):
    if raw_vaccine_tweets["user_location"][i] not in countries:
        raw_vaccine_tweets["user_location"][i] = None

Get coordinates for every country:

geolocator = Nominatim(user_agent="CovaxAnalytica")
location_coordinates = {}

for country in countries:
    location = geolocator.geocode(country)
    try:
        location_coordinates[country] = [location.latitude, location.longitude]
    except:
        location_coordinates[country] = None

location_coordinates

Map country name to coordinates:

raw_vaccine_tweets["coordinates"] = None
for i in range(len(raw_vaccine_tweets)):
     if raw_vaccine_tweets["user_location"][i] in countries:
        raw_vaccine_tweets["coordinates"][i] = location_coordinates[raw_vaccine_tweets["user_location"][i].lower()]

raw_vaccine_tweets

add the coordinates from the geo column:

for i in range(len(raw_vaccine_tweets)):
    if raw_vaccine_tweets.geo[i] != None:
        raw_vaccine_tweets.coordinates[i] = raw_vaccine_tweets.geo[i]["coordinates"]

renaming all countries in standardized way using "reverse_geocode":

for i in range(len(raw_vaccine_tweets)):
    if raw_vaccine_tweets.coordinates[i] != None:
        raw_vaccine_tweets.user_location[i] = reverse_geocode.search(tuple([raw_vaccine_tweets.coordinates[i],(1,1)]))[0]["country"]


raw_vaccine_tweets["user_location"].unique()

# Exporting dataset

raw_vaccine_tweets.to_csv("../data/interim/cleaned_vaccine_tweets.csv")

---
