==> notebooks/03_AM_Sentiment_Analysis_Vaccine_Tweets.ipynb <==
import pandas as pd
import numpy

# Sentiment analysis

- Preprocessing
    - remove URLs
    - remove Stopwords
    - Lemmatization
- Sentiment Analysis with nltk

clean_vaccine_tweets = pd.read_csv("../data/interim/cleaned_vaccine_tweets.csv", index_col=0)

clean_vaccine_tweets.head()

---

import re
import nltk
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from nltk.sentiment import SentimentIntensityAnalyzer

# NLP
## Preprocessing

clean_vaccine_tweets["corpus"] = ""

nltk.download('stopwords')
nltk.download('wordnet')
nltk.download('vader_lexicon')

Set lowercase, remove punctuation, remove url:
- "no","nor","not" removed from stopwords because they may be relevant for sentiment
- add "vaccine" to stopwords because it has little information value

def clean_dataset(dataset):
    for i in range(0, len(dataset)):
        #Tokenize and set words to lowercase 
        review = dataset["full_text"][i]
        review = ' '.join(re.sub("(@[A-Za-z0-9]+)|([^0-9A-Za-z \t])|(\w+:\/\/\S+)"," ",review).split())
        review = re.sub("[^a-zA-Z]", " ", review)
        review = review.lower()
        review = review.split()

        #stopwords: 
        all_stopwords = [word for word in stopwords.words("english") if word not in ["no","nor","not"]]
        all_stopwords.extend(["a","b","c","d","e","f","g","h","i","j","k","l","m","n","o","p","q","r","s","t","u","v","w","x","y","z","&amp","amp","amp;","&amp;","nhs","vaccine","covidvaccine"])
        #lemmatization:
        lemma = nltk.wordnet.WordNetLemmatizer()
        review = " ".join([lemma.lemmatize(word) for word in review if word not in set(all_stopwords)])    
        dataset["corpus"][i] = review


clean_dataset(clean_vaccine_tweets)

clean_vaccine_tweets.head()

## Sentiment Analysis with nltk

clean_vaccine_tweets["sentiment"] = dict
clean_vaccine_tweets["sentiment_compound"] = 0.0

Determine sentiment via NLTK.Sentiment:
- Values range from [-1, 1]
- -1 is negative, 0 is neutral, 1 is positive

def sentiment_score(dataset):
    sia = SentimentIntensityAnalyzer()
    for i in range(len(dataset)):
        dataset["sentiment"][i] = sia.polarity_scores(dataset["corpus"][i])
        dataset["sentiment_compound"][i] = dataset["sentiment"][i]["compound"]

sentiment_score(clean_vaccine_tweets)

clean_vaccine_tweets

Sort dataset by sentiment and create a sub-dataset with location only:

clean_vaccine_tweets = clean_vaccine_tweets.sort_values("sentiment_compound", ascending=False, ignore_index=True)

geo_vaccine_tweets = clean_vaccine_tweets[clean_vaccine_tweets["user_location"].isnull() == False].reset_index(drop=True)

---

Export:

clean_vaccine_tweets.to_csv("../data/processed/vaccine_tweets_with_sentiment.csv")

geo_vaccine_tweets.to_csv("../data/processed/geo_vaccine_tweets_with_sentiment.csv")
