==> notebooks/07_LB_Transform_Tweets.ipynb <==
# Overall Twitter Dataset including tweets regarding Covid-19 from November 2020 - June 2021

## Imports

import pandas as pd

## Reading Data

#Dataset including Tweet ID and Sentiment
ids = pd.read_csv('../data/raw/Corona_Combined_Nov2020-June2021.csv', header=None, encoding="utf-8", error_bad_lines=False)

#Hydrated Dataset of the above
hydrated = pd.read_csv('../data/raw/Hydrated_Tweets_Nov2020-Jun2021.csv')

## Adjusting the Dataset adding the Sentiment Score to each Entry

ids = ids.rename(columns={0:"id", 1:"sentiment"})
ids.id = ids.id.astype('str')
hydrated.id=hydrated.id.astype('str')
hydrated = hydrated.merge(ids, on="id", how="left")

### Splitting coordinates in latitude and longitude## 

df = hydrated
df[['longitude','latitude']] = df.coordinates.str.split(",",expand=True)

Export Dataset as CSV

df.to_csv("../data/interim/Overall_Map_data.csv")

# Using secondly hydrated Dataset with advanced Geo information

## Dataset changes:
<br>
- Adding Sentiment Score to each Entry<br>
- Add new Column 'date'<br>
- Break down the timestamp to date<br>
- Drop unused Columns<br>

#Read hydrated Dataset including more detailed Geo information
data = pd.read_csv('../data/interim/Hydrated_Tweets_with_countries.csv')

ids = ids.rename(columns={0:"id", 1:"sentiment"})
ids.id = ids.id.astype('float64')
data.id = data.id.astype('float64')
data = data.merge(ids, on="id", how="left")

data['date'] = ''
data['date'] = pd.to_datetime(data['created_at']).dt.strftime('%Y-%m-%d')

data.drop(columns=['Unnamed: 0', 'user', 'place', 'contributors', 'is_quote_status',
       'retweet_count', 'favorite_count', 'favorited', 'retweeted',
       'possibly_sensitive', 'lang', 'quoted_status_id','quoted_status_id_str', 'quoted_status_permalink', 'quoted_status',
       'extended_entities', 'withheld_in_countries'], inplace=True)

### Drop all entries with empty date + adding month and year

data = data.drop(data[data['date'].isnull()].index)
data['month'] = ''
data['year'] = ''

#Switched because american date format
data['month'] = pd.DatetimeIndex(data['date']).month
data['year'] = pd.DatetimeIndex(data['date']).year

### Creating Column that includes Tweets per day for each country

data['tweet_amount'] = ''
data.groupby(['date','country'])['id'].count().reset_index(name="tweet_amount")

Export Dataset as CSV

data.to_csv("../data/interim/Overall_data.csv")

## Using Vaccine Dataset

vaccine_data = pd.read_csv('../data/raw/country_vaccinations.csv')

### Dropping Dates which are not included in the Overall Dataset

vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-14'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-15'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-16'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-17'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-18'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-19'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-20'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-21'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-22'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-23'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-24'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-25'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-26'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-27'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-28'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-29'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-06-30'].index)
vaccine_data = vaccine_data.drop(vaccine_data[vaccine_data['date'] == '2021-07-01'].index)

Export Dataset as CSV

vaccine_data.to_csv("../data/interim/Vaccine_data.csv")
